+++ src/DependencyInjection/PentatrionViteExtension.php
@@ -27,7 +27,7 @@ class PentatrionViteExtension extends Extension
 
     public function load(array $configs, ContainerBuilder $container): void
     {
-        $loader = new YamlFileLoader($container, new FileLocator(__DIR__.'/../Resources/config'));
+        $loader = new YamlFileLoader($container, new FileLocator(\dirname(__DIR__).'/Resources/config'));
         $loader->load('services.yaml');
 
         $config = $this->processConfiguration(