+++ tests/Asset/TagRendererTest.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Pentatrion\ViteBundle\Tests\Asset;
+
+use PHPUnit\Framework\TestCase;
+
+class TagRendererTest extends TestCase
+{
+    public function testRenderTagBasic()
+    {
+        $this->assertEquals(1, 1);
+    }
+}